ITERATORS FOR Java Script
JavaScript iterators are objects that define a sequence and potentially a return value upon termination. They are used to traverse over a list or collection, allowing you to access each element in a controlled manner. Iterators are returned by the `Symbol.iterator()` method, which is available on iterable objects like arrays. The iterator protocol provides the `next()` method to access each element of the sequence.  Iterators and generator functions are advanced concepts in JavaScript that allow for efficient and customizable looping through data structures.
--> Symbol.iterator()
--> next()