LOOPS Java Script
Loops in JavaScript are a powerful tool for repeating a block of code multiple times. They are essential for efficiently handling repetitive tasks.  There are several types of loops in JavaScript, including for, while, do while, and for-in loops. Each loop has its own syntax and purpose. For example, the **for loop** repeats until a specified condition evaluates to false, while the while loop loops through a block of code as long as a specified condition is true. 
--> For Loops
--> While